Reworked rights display (again)
Refactored Blossom’s metadata bar to be flexible and reflow if needed,
which allows us to place the rights info inside it. Modified Rights
Module to return a copyright symbol for “all rights reserved” content.
Switched a couple of style rules to REMs to make the layout easier to
understand and edit.