Modified ReCAPTCHA and admin behaviour
ReCAPTCHA module now checks to make sure an API key is available before
proceeding; prevents the library from throwing a site-blocking error.

Admin previews are now activated by clicking the label for the field to
be previewed. Magnifier icon is appended to the label to indicate the
field can be previewed, and default focusing behaviour is prevented.

Sorted out the domain for locale strings in the admin theme’s
JavaScript.

Made a small CSS tweak to separate adjacent confirmation buttons in the
admin area.