Comments: chronological ordering and replies
Getting comments to display in chronological order, with the new
comment form at the bottom and without ajax gremlins.

Now using a shim to determine placement of ajax-loaded comments.

Refactored the reply function so it works properly with clean URLs, and
got Blossom to support it. Full reply functionality requires JavaScript
support.