Added first/final link generation to Paginator
Generate links to the first and final page in a paginated set.

Paginator’s link functions modified to `return` instead of `echo` –
this enables further processing of the returned string in Twig.

Admin theme updated: paginated sets will now generate a link to the
final page within the “Page x of y” text. Should be a help to users
with many posts.

Generated fresh .pot files.