Using IP-bound token to secure ajax
Benefits: not divulging the site’s hashkey if the JavaScript is linked
to by a bad actor in an CSRF scenario, and the token is bound to the
logged-in requestor’s IP address so it is of limited usefulness to any
third party. Also fixed protocol reporting in various error headers.