Likes module: implemented "sticky" likes for guests
Like status is stored in the visitor’s session. For logged-in users we
can reconstruct the session object if it’s destroyed; for guest
visitors the likes are sticky: remembered until session destruction
(cookie expiry).

When guest visitors unlike a post, the record is removed from their
session but not the database (to prevent deletion of other visitors’
likes on a shared IP).

If we discover a like in the database from a guest visitor’s IP but
their session has no record of it, we note it with a null to avoid the
possibility of attempting to add a duplicate hash to the database.

tl;dr: for logged-in users likes are persistent and reversible, for
guest visitors likes are sticky but not persistent and not truly
reversible.