Post and Page: changes to fallbacks and sanitisation
Posts and Pages will try to sanitise a $_POST['slug'] value to use as
clean, failing that they will use a random char string for additions
and the existing clean value for updates. If you provide these
functions with clean and URL values you are responsible for making sure
the values are web-safe and unique. Doing clones before we start work.