Admin: fixes to post_fields.twig and page_fields.twig
The slug attribute is actually a synonym for the unique URL attribute,
not the non-unique clean attribute. We should be presenting the clean
value here even though to field is named “Slug” so that, if no changes
are made, both clean and URL values remain unaltered when the post is
updated.