Route: more strict test for single param actions in __construct()
Only add arg[0] to the try list if the route action remains
undetermined after parsing by the controller and if arg[0] is not
empty. Validity of the argument will be tested in init() and an error
thrown if it is invalid, same as would happen if “?action=foo-bar” was
requested without clean URLs.