Theme: updated cookie notification message
Modified language to comply with amendments to ePrivacy regulations.
Out of the box, Chyrp Lite will only use cookies for user
authentication, which is a “strictly necessary” use case. This
notification is no longer required (or adequate) to ensure compliance;
instead it is intended as an assurance to the visitor that their data
will not be misused by the site.