Helpers: changes to email() and correspond()
Don’t need the old workaround for PHP < 5.3; made the action filter
optional so the subject and message can be supplied directly by the
caller if that’s more efficient; added a new “send_correspondence”
trigger call so that modules can correspond with the user by
alternative medium, e.g. send the user an SMS instead of an email.