Route: changes to custom() and current()
Don't try to process custom routes for any controller other than MainController, since they are designed to work with it exclusively.

If Route is initiated without a controller, return null and let the caller handle it. This fixes a problem with publishing scheduled posts when the Read More module is enabled.