Cacher module: generate a hint for reverse proxies
The hint uniquely identifies a combination of visitor group and route controller, e.g. a visitor belonging to the Guest group visiting a page served by MainController. A reverse proxy server could choose to cache all responses with a certain hint value. To serve cached responses only to guests, a module would need to be created to respond to the triggers "runtime", "user_logged_in", and "user_logged_out". The module would set a companion cookie on these triggers that describes the user's login status for the proxy server to inspect on the next request.