Trigger: added decide() and changed call behaviour
The decide() function will try to ensure something meaningful is returned. It has special concatenation behaviour if all calls return a string; this is to enable trigger calls from Twig to return content for integration into a template. Change necessary because Twig 4.x will no longer output templates using an immediate echo, so an echo made within a trigger responder will not be integrated at the proper place in the template.