Tags module: better escaping fix
Problem here is not related to string escaping per se, but specifically  that LIKE strings allow for % and _ to be escaped and made literal, and the default escape char is usually backslash. We've added the clause in QueryBuilder to be certain of the character, and now we can escape literal backslashes. See #257.